% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Ask}
\alias{Ask}
\title{Create a question}
\usage{
Ask(name, type, label, ...)
}
\arguments{
\item{name}{Unique question name.}

\item{type}{The type of the question.  One of 'Text()', 'Integer()',
'Decimal()', 'Range()', 'Geopoint()', 'Date()', 'Time()',
'DateTime()', 'SelectOne(...)', and 'SelectMultiple(...)'.}

\item{label}{A string that states the question that we want to
display in the survey.}

\item{...}{Other named arguments, e.g., 'hint', 'constraint',
'required', 'choice_list'.}
}
\description{
This function creates a question.  It should be called
in the arguments passed to the function 'Survey'.
}
