% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Choice}
\alias{Choice}
\title{Create an item in a choice list}
\usage{
Choice(value, label, ...)
}
\description{
This function creates a choice item.  It should be
called in the arguments passed to the function 'ChoiceList'.
}
