% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ChoiceList}
\alias{ChoiceList}
\title{Create a choice list}
\usage{
ChoiceList(name, ...)
}
\arguments{
\item{name}{Unique choice list name.}

\item{...}{Choices created using the function 'Choice'.}
}
\description{
This function creates a choice list that can be
referenced by questions of type 'SelectOne' or 'SelectMultiple'.
It should be called in the arguments passed to the function
'Survey'.
}
