% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Range}
\alias{Range}
\title{Specify the "range" type for a question created with 'Ask'}
\usage{
Range()
}
\description{
Specify the "range" type for a question created with 'Ask'
}
\examples{
Ask(
  tomato_intensity,
  Range(),
  "How much do you like tomatoes?",
  parameters = "start=1 end=5 step=1"
)
}
