% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SelectMultiple}
\alias{SelectMultiple}
\title{Specify the "select_multiple" type for a question created with 'Ask'}
\usage{
SelectMultiple(list_name)
}
\description{
Specify the "select_multiple" type for a question created with 'Ask'
}
\examples{
Survey(
  form_id = cabbage_survey,
  form_version = auto,
  form_title = "Survey on Cabbages",
  # Define a choice list with various cabbages.
  #
  ChoiceList(
    cabbage_species,
    Choice(1, "Bok choy"),
    Choice(2, "Broccoli"),
    Choice(3, "Brussels sprouts"),
    Choice(4, "Cauliflower"),
    Choice(5, "Choy sum"),
    Choice(6, "Kohlrabi"),
    Choice(7, "Napa cabbage"),
    Choice(8, "Rutabaga"),
    Choice(9, "Savoy cabbage"),
    Choice(10, "Turnip"),
    Choice(99, "Other")
  ),
  # Multiple-choice question about cabbages.
  #
  Ask(
    favorite_cabbage,
    SelectMultiple(cabbage_species),
    "Which are your favorite cabbages?"
  )
)
}
