% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SelectOne}
\alias{SelectOne}
\title{Specify the "select_one" type for a question created with 'Ask'}
\usage{
SelectOne(list_name)
}
\description{
Specify the "select_one" type for a question created with 'Ask'
}
\examples{
Ask(
  paprika,
  SelectOne(yes_no),
  "Have you consumed any paprika in the past 24 hours?"
)
}
