% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Survey}
\alias{Survey}
\title{Create a survey definition}
\usage{
Survey(form_id, form_version, form_title, ...)
}
\arguments{
\item{form_id}{Form identifier.}

\item{form_version}{Form version.  It must either be a positive
integer or the phrase "auto".  If it is set to "auto", an
automatic version number is generated based on the current date
and time in UTC.}

\item{form_title}{Form title.}

\item{...}{Form contents created using the functions 'ChoiceList',
'Ask', 'Note', 'Calculate', 'CalculateHere', 'When', 'IfElse',
'Group', 'TimedGroup', and 'Repeat'.}
}
\value{
A list object that can be processed by the function
'write_xlsform'.
}
\description{
This function creates a survey definition.  It can be
combined with the functions 'ChoiceList', 'Ask', 'Note',
'Calculate', 'CalculateHere', 'When', 'IfElse', 'Group',
'TimedGroup', and 'Repeat' in order to specify the survey.
}
\examples{
form_definition <- Survey(
  form_id = foo_v1,
  form_version = auto,
  form_title = "Foo v1",
  ChoiceList(
    yes_no,
    Choice(1, "Yes"),
    Choice(0, "No")
  ),
  Ask(name, Text(), "First and last name:"),
  Ask(
    vegetables,
    SelectOne(yes_no),
    "Have you eaten your vegetables?"
  ),
  When(
    vegetables == 1,
    Ask(
      vegetables_red,
      SelectOne(yes_no),
      "Have you eaten any red vegetables?",
      hint = "Tomatoes and peppers are technically fruits."
    ),
    When(
      vegetables_red == 1,
      Ask(
        vegetables_red_count,
        Integer(),
        "How many kinds of red vegetables have you eaten?",
        constraint = . > 0 & . < 100
      )
    )
  )
)
}
