% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Text}
\alias{Text}
\title{Specify the "text" type for a question created with 'Ask'}
\usage{
Text()
}
\description{
Specify the "text" type for a question created with 'Ask'
}
\examples{
Ask(name_first, Text(), "What is your first name?")
}
