% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{to_xlsform_string}
\alias{to_xlsform_string}
\title{String literals in XLSForm are written with single quotes.  This
function escapes single quotes in the input and puts single quotes at
the beginning and the end.}
\usage{
to_xlsform_string(str)
}
\arguments{
\item{str}{The input string to escape.}
}
\value{
\value{A string that contains an XLSForm string literal.}
}
\description{
String literals in XLSForm are written with single quotes.  This
function escapes single quotes in the input and puts single quotes at
the beginning and the end.
}
