% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform.R
\name{write_xlsform}
\alias{write_xlsform}
\title{Write a survey to an Excel file in XLSForm format}
\usage{
write_xlsform(obj, filename)
}
\arguments{
\item{obj}{A survey object created with the 'Survey' function.}

\item{filename}{A string containing the output filename.}
}
\description{
Write a survey to an Excel file in XLSForm format
}
\examples{
form_definition <- Survey(
  form_id = foo_v1,
  form_version = auto,
  form_title = "Foo v1",
  ChoiceList(
    yes_no,
    Choice(1, "Yes"),
    Choice(0, "No")
  ),
  Ask(name, Text(), "First and last name:"),
  Ask(zucchini, SelectOne(yes_no), "Have you eaten zucchini squash?")
)

\dontshow{
.old_wd <- setwd(tempdir())
}
write_xlsform(form_definition, "foo_v1.xlsx")
\dontshow{
setwd(.old_wd)
}
}
