% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SelectOne}
\alias{SelectOne}
\title{Specify the "select_one" type for a question created with 'Ask'}
\usage{
SelectOne(list_name)
}
\description{
Specify the "select_one" type for a question created with 'Ask'
}
\examples{
Survey(
  form_id = juice_survey,
  form_version = auto,
  form_title = "Survey on Juices",
  ChoiceList(
    yes_no,
    Choice(1, "Yes"),
    Choice(0, "No")
  ),
  Ask(
    paprika,
    SelectOne(yes_no),
    "Have you consumed any cucumber juice in the past 24 hours?"
  )
)
}
